+++ CHANGELOG.rst
@@ -1,10 +1,13 @@
 Releases
 ========
 
-0.3.4 (unreleased)
+0.4.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added an option to disable "required/optional" strictness on structs.
+- Added ``to_primitive`` and ``from_primitive`` methods to generated types to
+  allow converting struct, union, and exception values to and from primitive
+  representations.
 
 
 0.3.3 (2015-10-05)